Move Scoreboard Email Sending into 'afterUpdate' Event
Also, fix template not parsing correctly
Also, fix wording in comments
Also, get secretary object correctly
Also, sendMail doesn't need to be promisified

It Works!
